rename: components 폴더의 공통 파일 병합
