+++ src/components/Board/CommentDetail/index.tsx
