+++ src/components/Board/Like/index.tsx
