+++ src/components/Board/Like/styles.module.scss
