+++ src/pages/Community/Detail/index.tsx
@@ -2,8 +2,8 @@ import { Link, Navigate, useNavigate, useParams } from 'react-router-dom';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
 import dayjs from 'dayjs';
 import Dompurify from 'dompurify';
-import Comments from '@/components/BoardComponents/Comments';
-import Like from '@/components/BoardComponents/Like';
+import Comments from '@/components/Board/Comments';
+import Like from '@/components/Board/Like';
 import { QueryKeys, restFetcher } from '@/queryClient';
 import NotFoundPage from '@/pages/NotFound';
 import { DeleteBoardAPI } from '@/apis/boards';
